Microshift on FCOS
If openshift/microshift#2603 will land, the first stage running the fake openvswitch3.1 build could be dropped.

+++ overlay.d/01-common/usr/share/zsh/30zinitrc.zsh
@@ -56,7 +56,7 @@ exists_or_install_latest_gh_release operator-framework operator-sdk "operator-sd
 
 # symbolic link for kubectl -> oc
 #
-ln -s "${LOCAL_BIN_PATH}/oc" "${LOCAL_BIN_PATH}/kubectl"
+[ -f "${LOCAL_BIN_PATH}"/kubectl ] || ln -s "${LOCAL_BIN_PATH}/oc" "${LOCAL_BIN_PATH}/kubectl"
 
 [ -f "${LOCAL_BIN_PATH}"/helm ] || {
   export HELM_INSTALL_DIR="${LOCAL_BIN_PATH}"
